Added nestedAttr filter component.
1. Added nested attribute filter component to stop using same code multiple times.
