+++ includes/automatic-translate/src/component/storeSourceString/index.js
@@ -1,3 +1,4 @@
+import FilterBlockNestedAttr from "../FilterNestedAttr";
 const { dispatch } = wp.data;
 
 let contentIndex = 0;
@@ -44,59 +45,11 @@ const filterTranslateAttr = (blockId, blockAttr, filterAttr) => {
                 dispatch('block-atfp/translate').contentSaveSource(filterKey, blockAttrContent, contentIndex);
                 contentIndex++;
             }
-        }
 
-        if (Object.getPrototypeOf(filterAttrObj) === Array.prototype) {
-            childAttrArray(idArr, filterAttrObj);
-        } else if (Object.getPrototypeOf(filterAttrObj) === Object.prototype) {
-            childAttr(idArr, filterAttrObj);
-        }
-    }
-
-    /**
-     * Iterates over the keys of the filter object and calls saveTranslatedAttr for each key.
-     * 
-     * @param {Array} idArr - The array of IDs.
-     * @param {Object} filterObj - The filter object to iterate over.
-     */
-    const childAttr = (idArr, filterObj) => {
-        Object.keys(filterObj).map(key => {
-            let filterObjType = filterObj;
-            filterObjType = filterObjType[key];
-            const newIdArr = new Array(...idArr, key);
-
-            saveTranslatedAttr(newIdArr, filterObjType);
-        });
-    }
-
-    /**
-     * Handles the attributes that are arrays and objects by recursively calling saveTranslatedAttr.
-     * 
-     * @param {Array} idArr - The array of IDs.
-     * @param {Array} attrFilter - The filter attribute array.
-     */
-    const childAttrArray = (idArr, attrFilter) => {
-        const newIdArr = new Array(...idArr);
-        let dynamicBlockAttr = blockAttr;
-
-        newIdArr.forEach(key => {
-            dynamicBlockAttr = dynamicBlockAttr[key];
-        });
-
-        if (Object.getPrototypeOf(dynamicBlockAttr) === Object.prototype) {
-            childAttr(idArr, attrFilter[0]);
             return;
         }
 
-        dynamicBlockAttr.forEach((_, index) => {
-            const nestedId = new Array();
-            newIdArr.forEach(key => {
-                nestedId.push(key);
-            });
-
-            nestedId.push(index);
-            saveTranslatedAttr(nestedId, attrFilter[0]);
-        });
+        FilterBlockNestedAttr(idArr,filterAttrObj,blockAttr,saveTranslatedAttr);
     }
 
     filterAttrArr.forEach(data => {
